+++ src/containers/RecordTicketContainer/index.ts
@@ -1,0 +1,1 @@
+export { default, getServerSideProps } from './RecordTicketContainer';
